+++ src/ethereum/cancun/fork_types.py
@@ -13,159 +13,23 @@ Types re-used throughout the specification, which are specific to Ethereum.
 """
 
 from dataclasses import dataclass
-from typing import Tuple, Union
 
 from .. import rlp
 from ..base_types import (
-    U64,
     U256,
     Bytes,
-    Bytes0,
-    Bytes8,
     Bytes20,
-    Bytes32,
     Bytes256,
     Uint,
     slotted_freezable,
 )
 from ..crypto.hash import Hash32, keccak256
-from ..exceptions import InvalidBlock
 
 Address = Bytes20
 Root = Hash32
 VersionedHash = Hash32
 
 Bloom = Bytes256
-
-TX_BASE_COST = 21000
-TX_DATA_COST_PER_NON_ZERO = 16
-TX_DATA_COST_PER_ZERO = 4
-TX_CREATE_COST = 32000
-TX_ACCESS_LIST_ADDRESS_COST = 2400
-TX_ACCESS_LIST_STORAGE_KEY_COST = 1900
-
-
-@slotted_freezable
-@dataclass
-class LegacyTransaction:
-    """
-    Atomic operation performed on the block chain.
-    """
-
-    nonce: U256
-    gas_price: Uint
-    gas: Uint
-    to: Union[Bytes0, Address]
-    value: U256
-    data: Bytes
-    v: U256
-    r: U256
-    s: U256
-
-
-@slotted_freezable
-@dataclass
-class AccessListTransaction:
-    """
-    The transaction type added in EIP-2930 to support access lists.
-    """
-
-    chain_id: U64
-    nonce: U256
-    gas_price: Uint
-    gas: Uint
-    to: Union[Bytes0, Address]
-    value: U256
-    data: Bytes
-    access_list: Tuple[Tuple[Address, Tuple[Bytes32, ...]], ...]
-    y_parity: U256
-    r: U256
-    s: U256
-
-
-@slotted_freezable
-@dataclass
-class FeeMarketTransaction:
-    """
-    The transaction type added in EIP-1559.
-    """
-
-    chain_id: U64
-    nonce: U256
-    max_priority_fee_per_gas: Uint
-    max_fee_per_gas: Uint
-    gas: Uint
-    to: Union[Bytes0, Address]
-    value: U256
-    data: Bytes
-    access_list: Tuple[Tuple[Address, Tuple[Bytes32, ...]], ...]
-    y_parity: U256
-    r: U256
-    s: U256
-
-
-@slotted_freezable
-@dataclass
-class BlobTransaction:
-    """
-    The transaction type added in EIP-4844.
-    """
-
-    chain_id: U64
-    nonce: U256
-    max_priority_fee_per_gas: Uint
-    max_fee_per_gas: Uint
-    gas: Uint
-    to: Address
-    value: U256
-    data: Bytes
-    access_list: Tuple[Tuple[Address, Tuple[Bytes32, ...]], ...]
-    max_fee_per_blob_gas: U256
-    blob_versioned_hashes: Tuple[VersionedHash, ...]
-    y_parity: U256
-    r: U256
-    s: U256
-
-
-Transaction = Union[
-    LegacyTransaction,
-    AccessListTransaction,
-    FeeMarketTransaction,
-    BlobTransaction,
-]
-
-
-def encode_transaction(tx: Transaction) -> Union[LegacyTransaction, Bytes]:
-    """
-    Encode a transaction. Needed because non-legacy transactions aren't RLP.
-    """
-    if isinstance(tx, LegacyTransaction):
-        return tx
-    elif isinstance(tx, AccessListTransaction):
-        return b"\x01" + rlp.encode(tx)
-    elif isinstance(tx, FeeMarketTransaction):
-        return b"\x02" + rlp.encode(tx)
-    elif isinstance(tx, BlobTransaction):
-        return b"\x03" + rlp.encode(tx)
-    else:
-        raise Exception(f"Unable to encode transaction of type {type(tx)}")
-
-
-def decode_transaction(tx: Union[LegacyTransaction, Bytes]) -> Transaction:
-    """
-    Decode a transaction. Needed because non-legacy transactions aren't RLP.
-    """
-    if isinstance(tx, Bytes):
-        if tx[0] == 1:
-            return rlp.decode_to(AccessListTransaction, tx[1:])
-        elif tx[0] == 2:
-            return rlp.decode_to(FeeMarketTransaction, tx[1:])
-        elif tx[0] == 3:
-            return rlp.decode_to(BlobTransaction, tx[1:])
-        else:
-            raise InvalidBlock
-    else:
-        return tx
 
 
 @slotted_freezable
@@ -202,83 +66,3 @@ def encode_account(raw_account_data: Account, storage_root: Bytes) -> Bytes:
             keccak256(raw_account_data.code),
         )
     )
-
-
-@slotted_freezable
-@dataclass
-class Withdrawal:
-    """
-    Withdrawals that have been validated on the consensus layer.
-    """
-
-    index: U64
-    validator_index: U64
-    address: Address
-    amount: U256
-
-
-@slotted_freezable
-@dataclass
-class Header:
-    """
-    Header portion of a block on the chain.
-    """
-
-    parent_hash: Hash32
-    ommers_hash: Hash32
-    coinbase: Address
-    state_root: Root
-    transactions_root: Root
-    receipt_root: Root
-    bloom: Bloom
-    difficulty: Uint
-    number: Uint
-    gas_limit: Uint
-    gas_used: Uint
-    timestamp: U256
-    extra_data: Bytes
-    prev_randao: Bytes32
-    nonce: Bytes8
-    base_fee_per_gas: Uint
-    withdrawals_root: Root
-    blob_gas_used: U64
-    excess_blob_gas: U64
-    parent_beacon_block_root: Root
-
-
-@slotted_freezable
-@dataclass
-class Block:
-    """
-    A complete block.
-    """
-
-    header: Header
-    transactions: Tuple[Union[Bytes, LegacyTransaction], ...]
-    ommers: Tuple[Header, ...]
-    withdrawals: Tuple[Withdrawal, ...]
-
-
-@slotted_freezable
-@dataclass
-class Log:
-    """
-    Data record produced during the execution of a transaction.
-    """
-
-    address: Address
-    topics: Tuple[Hash32, ...]
-    data: bytes
-
-
-@slotted_freezable
-@dataclass
-class Receipt:
-    """
-    Result of a transaction.
-    """
-
-    succeeded: bool
-    cumulative_gas_used: Uint
-    bloom: Bloom
-    logs: Tuple[Log, ...]
