+++ src/ethereum/cancun/state.py
@@ -22,7 +22,8 @@ from typing import Callable, Dict, Iterable, List, Optional, Set, Tuple
 from ethereum.base_types import U256, Bytes, Uint, modify
 from ethereum.utils.ensure import ensure
 
-from .fork_types import EMPTY_ACCOUNT, Account, Address, Root, Withdrawal
+from .blocks import Withdrawal
+from .fork_types import EMPTY_ACCOUNT, Account, Address, Root
 from .trie import EMPTY_TRIE_ROOT, Trie, copy_trie, root, trie_get, trie_set
